live_trading/requirements.txt
# CSP Strategy - Live Trading Dependencies
# Pin versions for reproducibility

alpaca-py>=0.43.0,<1.0
yfinance>=1.0,<2.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0
py-vollib>=1.0.1
python-dotenv>=1.0,<2.0
pytz>=2024.1
requests>=2.31,<3.0
lxml>=5.0,<6.0

# Cloud Run deployment
google-cloud-storage>=2.14.0,<3.0

# Testing (not installed in Docker image)
pytest-asyncio>=0.24.0,<1.0
